backend/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.12.2
gunicorn==20.1.0
Pillow==8.1.2
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
